requirements.txt
#algorand==0.1.0
autopep8==1.7.0
cffi==1.15.1
click==8.1.3
flake8==5.0.4
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
py-algorand-sdk==1.18.0
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pyflakes==2.5.0
PyNaCl==1.5.0
python-dotenv==0.21.0
toml==0.10.2
Werkzeug==2.2.2
